refactor(shared): extract execute_tool_call into shared.tool_execution
Agent-loop plugins diverged in how they updated tool call context. Use the stricter shared execution path that preserves tool-specific UserValves while still refreshing per-call dynamic context.
