+++ src/owui_ext/tools/sub_agent.py
@@ -24,11 +24,9 @@ Inspired by VS Code's runSubagent functionality, this tool was developed from sc
 """
 
 import asyncio
-import ast
 import json
 import logging
 import re
-import uuid
 from collections.abc import Mapping, Sequence
 from typing import Any, Callable, List, Literal, Optional, Type
 
@@ -48,9 +46,10 @@ from owui_ext.shared.prompt_utils import (
     merge_prompt_sections,
 )
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
-from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
+from owui_ext.shared.tool_event_metadata import CITATION_TOOLS
 from owui_ext.shared.tool_execution import (
     execute_direct_tool_call,
+    execute_tool_call,
     normalize_terminal_tools_result,
     process_tool_result,
 )
@@ -447,165 +446,6 @@ async def cleanup_mcp_clients(mcp_clients: dict) -> None:
             log.debug(f"[SubAgent] Error cleaning up MCP client: {e}")
 
 
-async def execute_tool_call(
-    tool_call: dict,
-    tools_dict: dict,
-    extra_params: dict,
-    event_emitter: Optional[Callable] = None,
-) -> dict:
-    """Execute a single tool call and return the result.
-
-    Args:
-        tool_call: The tool call dict with id, function.name, function.arguments
-        tools_dict: Dict of available tools {name: {callable, spec, ...}}
-        extra_params: Extra parameters to pass to tool functions
-        event_emitter: Optional event emitter for citation/source events
-
-    Returns:
-        Dict with tool_call_id and content (result)
-    """
-    if not isinstance(tool_call, dict):
-        return {
-            "tool_call_id": str(uuid.uuid4()),
-            "content": f"Malformed tool_call: expected dict, got {type(tool_call).__name__}",
-        }
-    tool_call_id = tool_call.get("id", str(uuid.uuid4()))
-    func = tool_call.get("function")
-    if not isinstance(func, dict):
-        return {
-            "tool_call_id": tool_call_id,
-            "content": f"Malformed tool_call: 'function' is {type(func).__name__}, not dict",
-        }
-    tool_function_name = func.get("name", "")
-    tool_args_raw = func.get("arguments", "{}")
-
-    # Parse arguments - handle both JSON string and pre-parsed dict/list
-    tool_function_params: dict = {}
-    if isinstance(tool_args_raw, dict):
-        tool_function_params = tool_args_raw
-    elif isinstance(tool_args_raw, str):
-        try:
-            tool_function_params = ast.literal_eval(tool_args_raw)
-        except Exception:
-            try:
-                tool_function_params = json.loads(tool_args_raw)
-            except Exception as e:
-                log.error(f"Error parsing tool call arguments: {tool_args_raw} - {e}")
-                return {
-                    "tool_call_id": tool_call_id,
-                    "content": f"Error parsing arguments: {e}",
-                }
-    if not isinstance(tool_function_params, dict):
-        tool_function_params = {}
-
-    # Execute the tool
-    tool_result = None
-    tool_result_files: list[dict] = []
-    tool_result_embeds: list[Any] = []
-    emit_terminal_event = False
-    if tool_function_name in tools_dict:
-        tool = tools_dict[tool_function_name]
-        spec = tool.get("spec", {})
-        direct_tool = bool(tool.get("direct", False))
-
-        try:
-            allowed_params = spec.get("parameters", {}).get("properties", {}).keys()
-            tool_function_params = {
-                k: v for k, v in tool_function_params.items() if k in allowed_params
-            }
-
-            if direct_tool:
-                tool_result = await execute_direct_tool_call(
-                    tool_function_name=tool_function_name,
-                    tool_function_params=tool_function_params,
-                    tool=tool,
-                    extra_params=extra_params,
-                )
-            else:
-                tool_function = tool["callable"]
-
-                # Update function with current context
-                # __event_emitter__ is included so parallel sub-agents can
-                # override it with a task-specific indexed emitter.
-                from open_webui.utils.tools import get_updated_tool_function
-
-                tool_function = await maybe_await(get_updated_tool_function(
-                    function=tool_function,
-                    extra_params={
-                        "__messages__": extra_params.get("__messages__", []),
-                        "__files__": extra_params.get("__files__", []),
-                        "__event_emitter__": extra_params.get("__event_emitter__"),
-                        "__event_call__": extra_params.get("__event_call__"),
-                    },
-                ))
-
-                tool_result = await tool_function(**tool_function_params)
-
-            # Handle OpenAPI/external tool results that return (data, headers) tuple
-            # Headers (CIMultiDictProxy) are not JSON-serializable
-            tool_type = tool.get("type", "")
-            tool_result, tool_result_files, tool_result_embeds = await process_tool_result(
-                tool_function_name=tool_function_name,
-                tool_type=tool_type,
-                tool_result=tool_result,
-                direct_tool=direct_tool,
-                request=extra_params.get("__request__"),
-                metadata=extra_params.get("__metadata__"),
-                user=extra_params.get("__user__"),
-            )
-            emit_terminal_event = True
-
-        except Exception as e:
-            log.exception(f"Error executing tool {tool_function_name}: {e}")
-            tool_result = f"Error: {e}"
-    else:
-        tool_result = f"Tool '{tool_function_name}' not found"
-
-    if emit_terminal_event:
-        await emit_terminal_tool_event(
-            tool_function_name=tool_function_name,
-            tool_function_params=tool_function_params,
-            tool_result=tool_result,
-            event_emitter=event_emitter,
-        )
-        if event_emitter and tool_result_files:
-            await event_emitter({"type": "files", "data": {"files": tool_result_files}})
-        if event_emitter and tool_result_embeds:
-            await event_emitter({"type": "embeds", "data": {"embeds": tool_result_embeds}})
-
-    # Process result to string
-    if tool_result is None:
-        tool_result = ""
-    elif not isinstance(tool_result, str):
-        try:
-            tool_result = json.dumps(tool_result, ensure_ascii=False, default=str)
-        except Exception:
-            tool_result = str(tool_result)
-
-    # Extract and emit citation sources for tools that generate them
-    # This enables proper source attribution in the UI (e.g., web search results, KB documents)
-    if event_emitter and tool_result and tool_function_name in CITATION_TOOLS:
-        try:
-            from open_webui.utils.middleware import get_citation_source_from_tool_result
-
-            tool_id = tools_dict.get(tool_function_name, {}).get("tool_id", "")
-            citation_sources = get_citation_source_from_tool_result(
-                tool_name=tool_function_name,
-                tool_params=tool_function_params,
-                tool_result=tool_result,
-                tool_id=tool_id,
-            )
-            for source in citation_sources:
-                await event_emitter({"type": "source", "data": source})
-        except Exception as e:
-            log.warning(f"Error extracting citation sources from {tool_function_name}: {e}")
-
-    return {
-        "tool_call_id": tool_call_id,
-        "content": tool_result,
-    }
-
-
 async def apply_inlet_filters_if_enabled(
     apply_inlet_filters: bool,
     request: Request,
@@ -934,6 +774,8 @@ async def run_sub_agent_loop(
                         "__messages__": current_messages,
                     },
                     event_emitter=event_emitter,
+                    emit_terminal_event_fn=emit_terminal_tool_event,
+                    citation_tools=CITATION_TOOLS,
                 )
 
                 # Emit status with tool result
